Fix ByteArray and nullable primitives factorization to make bomb tests pass
