+++ qbit-fss/build.gradle
@@ -25,6 +25,7 @@ kotlin {
             dependencies {
                 implementation 'org.jetbrains.kotlin:kotlin-test-common'
                 implementation 'org.jetbrains.kotlin:kotlin-test-annotations-common'
+                implementation project(':qbit-test-fixtures')
             }
         }
         jvmTest {
