+++ qbit-http-storages/build.gradle
@@ -31,6 +31,7 @@ kotlin {
             dependencies {
                 implementation 'org.jetbrains.kotlin:kotlin-test-common'
                 implementation 'org.jetbrains.kotlin:kotlin-test-annotations-common'
+                implementation project(':qbit-test-fixtures')
             }
         }
         jvmMain {
